Refactor and add shape preview
